+++ src/extension.ts
@@ -113,6 +113,6 @@ const debounceShowDraftTrigger = () => {
   return (fnc: any, time: number) => {
     const functionCall = (...args: any[]) => fnc.apply(args);
     clearTimeout(timeout);
-    timeout = setTimeout(functionCall, time);
+    timeout = setTimeout(functionCall, time) as any;
   };
 };